feature/지도-좌표표시 map.js 분리
