+++ src/main/resources/templates/home.html
@@ -96,7 +96,7 @@
 </section>
 <div layout:fragment="scripts">
     <script th:replace="~{map/map-picker-modal :: mapPickerModalJs}"></script>
-    <script src="/home/js/map.js"></script>
+    <script type="module" src="/home/js/map.js"></script>
 </div>
 </body>
 </html>
